Allow user to login with app provided native login screen.
