+++ libs/SalesforceSDKCore/SalesforceSDKCore/Classes/UserAccount/SFUserAccountManager.h
@@ -389,6 +389,10 @@ NS_SWIFT_NAME(UserAccountManager)
  */
 - (SFUserAccount*)createUserAccount:(SFOAuthCredentials *)credentials NS_SWIFT_NAME(createUserAccount(with:));
 
+/** Create an account when necessary using token endpoint response data.
+  @param data The token endpoint response to use.
+ */
+- (void)createNativeUserAccount:(NSData *)data NS_SWIFT_NAME(createNativeUserAccount(with:));
 
 /** Allows you to look up the user account associated with a given user identity.
  @param userIdentity The user identity of the user account to be looked up
